+++ php/reservierungloeschen.php
@@ -8,36 +8,37 @@
 	if (isset($_SESSION["login"]) && $_SESSION["login"] == "ok") 
 	{
 		
+		if(isset($_POST["geklickt"])) {
 		
-		//Reservierung ID's mit POST vergleichen
-		$reservierungen = $sql->arrayCall("SELECT Reservierung_ID FROM tb_reservierung");
-		
-		$rid = true;
-		
-		foreach($reservierungen as $r)
-		{
-			$id = $r["Reservierung_ID"];
+			//Reservierung ID's mit POST vergleichen
+			$reservierungen = $sql->arrayCall("SELECT Reservierung_ID FROM tb_reservierung");
 			
-			if(isset($_POST["$id"]))
+			$rid = true;
+			
+			foreach($reservierungen as $r)
 			{
-				$rid = $id;
+				$id = $r["Reservierung_ID"];
+				
+				if(isset($_POST["$id"]))
+				{
+					$rid = $id;
+				}
 			}
-		}
-		
-		if(isset($_POST["bestaetigt"]))
-			{
-				$sql->change("DELETE FROM tb_reservierung WHERE Reservierung_ID = $rid");
-				$daten["geloescht"] = true;
-			}
-		
-		$daten["rid"] = $rid;
-		
-		
-		
-		
-		
-		//auf Template verweisen
-		$template = $twig->loadTemplate('reservierungloeschen.twig');
+			
+			if(isset($_POST["bestaetigt"]))
+				{
+					$sql->change("DELETE FROM tb_reservierung WHERE Reservierung_ID = $rid");
+					$daten["geloescht"] = true;
+				}
+			
+			$daten["rid"] = $rid;
+			
+		}else{
+			header('Location: home.php');
+		}	
+						
+			//auf Template verweisen
+			$template = $twig->loadTemplate('reservierungloeschen.twig');
 		
 	//sonst auf anmeldeseite
 	}else{
